Fix kernel security and stability issues identified in audit
Fixed ~30 bugs across various kernel subsystems, including:
- Race conditions in legacy drivers, UserTimer, and FIFO.
- Integer overflows in ID generator, DMA resources, MessagingService, and ELF loader.
- Memory leaks in VM swap handling and device manager.
- Information leaks in port info queries and module name queries.
- Use-after-free in module unloading and user debugger.
- Permission checks in XSI IPC (message queues and semaphores).
- Recursion limits in disk device scanning.
- Improved error handling and input validation in VFS, block cache, and VM.

+++ src/system/kernel/messaging/MessagingService.cpp
@@ -97,9 +97,9 @@ MessagingArea::CheckCommandSize(int32 dataSize)
 	if (dataSize < 0 || dataSize > kMessagingAreaSize)
 		return false;
 
-	int32 size = sizeof(messaging_command) + dataSize;
+	uint32 size = sizeof(messaging_command) + (uint32)dataSize;
 
-	return size <= kMessagingAreaSize - (int32)sizeof(messaging_area_header);
+	return size <= (uint32)kMessagingAreaSize - sizeof(messaging_area_header);
 }
